+++ R/zzz.R
@@ -10,7 +10,6 @@
   make_rule_fit()
 
   s3_register("generics::tunable", "cubist_rules")
-  s3_register("generics::tunable", "rule_fit")
   s3_register("generics::tunable", "C5_rules")
 
   s3_register("generics::tidy", "cubist")